01-getting-started/02-integration-mcp-tools/requirements.txt
uv
strands-agents
strands-agents-tools
mcp
pip-system-certs
certifi